mapreduce()

For this function, there are three versions available - tmapreduce, which parallelises only the mapping part (only function f), maptreduce, which parallelises only the reduction part (only function op) and tmaptreduce, which parallelises both.

ThreadedIterables.tmapreduce — Function

tmapreduce(f::Function, op::Function, itr::AbstractArray)

Multi-threaded version of mapreduce(f, op, itr). Only the mapping part is parallelised.

ThreadedIterables.maptreduce — Function

maptreduce(f::Function, op::Function, itr::AbstractArray)

Multi-threaded version of mapreduce(f, op, itr). Only the reduction part is parallelised. Note that function op must not change type!

ThreadedIterables.tmaptreduce — Function

tmaptreduce(f::Function, op::Function, itr::AbstractArray)

Multi-threaded version of mapreduce(f, op, itr). Both the mapping and the reduction part is parallelised. Note that function op must not change type!